+++ docs/.vuepress/config.js
@@ -44,7 +44,6 @@ module.exports = config({
         }
     },
     themeConfig: {
-        breadcrumb: false,
         mdEnhance: {
           align: true,
           tasklist: true,
@@ -101,19 +100,29 @@ module.exports = config({
                     }
                 ],
                 sidebar: {
-                    '/en/guide/': [{
-                        title: 'Getting started',
-                        collapsable: false,
-                        children: [
-                            '',
-                            'install',
-                            'getstarted',
-                            'feature',
-                            'whatsnew',
-                            'history',
-                            'faq'
-                        ]
-                    }]
+                    '/en/guide/': [
+                        "introduce",
+                        {
+                                  title: "Getting started",
+                                  icon: "creativefill",
+                                  prefix: "getstarted/",
+                                  collapsable: false,
+                                  children: [
+                                      "start",
+                                      "dataMapping",
+                                      "icon",
+                                      "otherConfig",
+                                      "projectEnv",
+                                      "README",
+                                      "stringGenerationStrategy"
+                                  ],
+                        },
+                        "install",
+                        "feature",
+                        "whatsnew",
+                        "history",
+                        "faq",
+                    ]
                 }
             },
             '/': {
@@ -145,20 +154,30 @@ module.exports = config({
                     }
                 ],
                 sidebar: {
-                    '/guide/': [{
-                        title: '快速入门',
-                        collapsable: false,
-                        children: [
-                            '',
-                            'install',
-                            'getstarted',
-                            'feature',
-                            'whatsnew',
-                            'history',
-                            'faq',
-                            'donate'
-                        ]
-                    }]
+                    '/guide/': [
+                        "introduce",
+                        {
+                                  title: "快速上手",
+                                  icon: "creativefill",
+                                  prefix: "getstarted/",
+                                  collapsable: false,
+                                  children: [
+                                      "start",
+                                      "dataMapping",
+                                      "icon",
+                                      "otherConfig",
+                                      "projectEnv",
+                                      "README",
+                                      "stringGenerationStrategy"
+                                  ],
+                        },
+                        "install",
+                        "feature",
+                        "whatsnew",
+                        "history",
+                        "faq",
+                        "donate"
+                    ],
                 }
             }
         }